# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/auth/init": {
            "post": {
                "description": "Creates the very first administrator (admin) account for the system using the provided email and password. This endpoint can only be called successfully once, when no admin accounts exist in the system yet.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Initialize first administrator account",
                "parameters": [
                    {
                        "description": "Admin initialization request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.InitAdminRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.InitAdminResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http_error.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Authenticates a user using their email and password credentials. On successful authentication, creates a session and sets an HTTP-only cookie for subsequent authenticated requests.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User authentication login",
                "parameters": [
                    {
                        "description": "Login request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http_error.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http_error.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "description": "Invalidates the current session by deleting it from the database and clearing the session cookie.",
                "tags": [
                    "Auth"
                ],
                "summary": "User logout",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http_error.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/session": {
            "get": {
                "description": "Returns the current active session details.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Get current session",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Session"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http_error.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.InitAdminRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "dto.InitAdminResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "dto.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "dto.LoginResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Login successful"
                }
            }
        },
...
        "http_error.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "models.Session": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "CreatedAt tracks when the record was created\nAutomatically set by GORM on create",
                    "type": "string"
                },
                "expiresAt": {
                    "description": "ExpiresAt is when this session becomes invalid",
                    "type": "string"
                },
                "id": {
                    "description": "ID is the primary key using UUID v4 format\nGORM tag ensures it's stored as uuid type in PostgreSQL with primary key constraint",
                    "type": "string"
                },
                "ipAddress": {
                    "description": "IPAddress stores the IP address where the session was created",
                    "type": "string"
                },
                "sessionToken": {
                    "description": "SessionToken is the unique token for this session (stored in cookie)",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "UpdatedAt tracks when the record was last modified\nAutomatically updated by GORM on save/update",
                    "type": "string"
                },
                "user": {
                    "description": "User relationship (many sessions belong to one user)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "userAgent": {
                    "description": "UserAgent stores the browser/user agent string",
                    "type": "string"
                },
                "userId": {
                    "description": "UserID is the foreign key to the user",
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "CreatedAt tracks when the record was created\nAutomatically set by GORM on create",
                    "type": "string"
                },
                "email": {
                    "description": "Email is the unique identifier for login\nMust be unique and not null",
                    "type": "string"
                },
                "id": {
                    "description": "ID is the primary key using UUID v4 format\nGORM tag ensures it's stored as uuid type in PostgreSQL with primary key constraint",
                    "type": "string"
                },
                "name": {
                    "description": "Name is the display name of the user",
                    "type": "string"
                },
                "role": {
                    "description": "Role determines user permissions (ADMIN or USER)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.UserRole"
                        }
                    ]
                },
                "updatedAt": {
                    "description": "UpdatedAt tracks when the record was last modified\nAutomatically updated by GORM on save/update",
                    "type": "string"
                }
            }
        },
        "models.UserRole": {
            "type": "string",
            "enum": [
                "ADMIN",
                "USER"
            ],
            "x-enum-varnames": [
                "RoleAdmin",
                "RoleUser"
            ]
        }
    }
}